backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI RAG System Backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # FastAPI & ASGI Server
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "pgvector>=0.3",
    # Migrations
    "alembic>=1.14",
    # AI
    "openai>=1.60",
    "tiktoken>=0.8",
    # Settings & Validation
    "pydantic-settings>=2.7",
    "email-validator>=2.2",
    # File upload & parsing
    "python-multipart>=0.0.18",
    "pypdf>=5.0",
    # Auth
    "PyJWT>=2.10",
    # Cache (token blacklist)
    "redis>=5.2",
    # Logging & Resilience
    "loguru>=0.7.3",
    "tenacity>=9.1",
    # HTTP client
    "httpx>=0.28",
    "bcrypt>=5.0.0",
]

[dependency-groups]
dev = [
    # Linting & Formatting
    "ruff>=0.14",
    # Type Checking
    "mypy>=1.14",
    "types-redis>=4.6",
    # Testing
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "pytest-xdist>=3.8",
    "pytest-timeout>=2.3",
    "httpx>=0.28",
    # Pre-commit
    "pre-commit>=4.5",
    "aiosqlite>=0.22.1",
]

[tool.ruff]
target-version = "py313"
line-length = 100
exclude = ["alembic/versions/*"]

[tool.ruff.lint]
select = [
    "F",       # Pyflakes
    "I",       # isort
    "N",       # pep8-naming
    "W",       # pycodestyle warnings
    "UP",      # pyupgrade
    "B",       # flake8-bugbear
    "C4",      # flake8-comprehensions
    "SIM",     # flake8-simplify
    "S",       # flake8-bandit (security)
    "T20",     # flake8-print
    "RUF",     # Ruff specific
    "PLC0415", # pylint - import at top level
]
ignore = [
    # Naming
    "N815",    # mixedCase in Schema
    "N805",    # Pydantic validator uses cls
    # Security - acceptable in context
    "S101",    # assert in tests
    "S104",    # bind all IPs in containers
    "S105",    # hardcoded password (false positives)
    "S106",    # hardcoded password argument
    "S110",    # try-except-pass
    "S113",    # httpx timeout=None
    "S301",    # serialization
    "S311",    # pseudo-random for non-crypto
    "S608",    # SQL injection (false positives)
    # Ruff - Chinese language support
    "RUF001",  # Ambiguous unicode in strings
    "RUF002",  # Ambiguous unicode in docstrings
    "RUF003",  # Ambiguous unicode in comments
    "RUF012",  # Mutable class attributes
    "RUF059",  # Unpacked variable never used
    # Simplify - readability preferences
    "SIM102",  # Nested if statements
    "SIM105",  # Use contextlib.suppress
    "SIM108",  # Use ternary operator
    "SIM110",  # Use any() instead of for loop
    "SIM115",  # Use context manager for file
    # Bugbear
    "B006",    # Mutable default argument
    "B008",    # Function call in default argument (FastAPI Depends pattern)
    "B905",    # zip without strict
    # Comprehensions
    "C416",    # Unnecessary list comprehension
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PLC0415"]

[tool.mypy]
python_version = "3.13"
warn_return_any = false
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = false
disallow_untyped_defs = false
check_untyped_defs = true
strict_equality = true
ignore_missing_imports = true
no_implicit_optional = false
show_error_codes = true
show_column_numbers = true
pretty = true

exclude = [
    "alembic/",
    "venv/",
    ".venv/",
]

[[tool.mypy.overrides]]
module = "*.tests.*"
ignore_errors = false
disallow_untyped_defs = false
check_untyped_defs = false

[tool.pytest.ini_options]
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = "-n auto --dist worksteal --timeout=30"
filterwarnings = [
    "ignore::DeprecationWarning",
]
